walk: restore the recursive iterator
The `ag/sys` branch leaves the new `WalkDir` implementation
unfinished. This brings back the existing path-based traversal without
using the new platform APIs yet.

This gives the later commits a working baseline.
