+++ src/tests/util.rs
@@ -1,19 +1,13 @@
 use std::env;
 use std::error;
-#[cfg(any(unix, windows))]
-use std::ffi::OsString;
 use std::fs::{self, File};
 use std::io;
 use std::path::{Path, PathBuf};
 use std::result;
 
-#[cfg(unix)]
-use crate::os::unix;
-#[cfg(windows)]
-use crate::os::windows;
 use crate::{DirEntry, Error};
 
-/// Skip the current test if the current environment doesn't support symlinks.
+/// Skip the current test if the environment cannot create symlinks.
 #[macro_export]
 macro_rules! skip_if_no_symlinks {
     () => {
@@ -87,116 +81,6 @@ impl RecursiveResults {
     }
 }
 
-/// The result of running a Unix directory iterator on a single directory.
-#[cfg(unix)]
-#[derive(Debug)]
-pub struct UnixResults {
-    ents: Vec<unix::DirEntry>,
-    errs: Vec<io::Error>,
-}
-
-#[cfg(unix)]
-impl UnixResults {
-    /// Return all of the errors encountered during traversal.
-    pub fn errs(&self) -> &[io::Error] {
-        &self.errs
-    }
-
-    /// Assert that no errors have occurred.
-    pub fn assert_no_errors(&self) {
-        assert!(
-            self.errs.is_empty(),
-            "expected to find no errors, but found: {:?}",
-            self.errs
-        );
-    }
-
-    /// Return all the successfully retrieved directory entries in the order
-    /// in which they were retrieved.
-    pub fn ents(&self) -> &[unix::DirEntry] {
-        &self.ents
-    }
-
-    /// Return all file names from all successfully retrieved directory
-    /// entries.
-    ///
-    /// This does not include file names that correspond to an error.
-    pub fn file_names(&self) -> Vec<OsString> {
-        self.ents.iter().map(|d| d.file_name_os().to_os_string()).collect()
-    }
-
-    /// Return all the successfully retrieved directory entries, sorted
-    /// lexicographically by their file name.
-    pub fn sorted_ents(&self) -> Vec<unix::DirEntry> {
-        let mut ents = self.ents.clone();
-        ents.sort_by(|e1, e2| e1.file_name_bytes().cmp(e2.file_name_bytes()));
-        ents
-    }
-
-    /// Return all file names from all successfully retrieved directory
-    /// entries, sorted lexicographically.
-    ///
-    /// This does not include file names that correspond to an error.
-    pub fn sorted_file_names(&self) -> Vec<OsString> {
-        self.sorted_ents().into_iter().map(|d| d.into_file_name_os()).collect()
-    }
-}
-
-/// The result of running a Windows directory iterator on a single directory.
-#[cfg(windows)]
-#[derive(Debug)]
-pub struct WindowsResults {
-    ents: Vec<windows::DirEntry>,
-    errs: Vec<io::Error>,
-}
-
-#[cfg(windows)]
-impl WindowsResults {
-    /// Return all of the errors encountered during traversal.
-    pub fn errs(&self) -> &[io::Error] {
-        &self.errs
-    }
-
-    /// Assert that no errors have occurred.
-    pub fn assert_no_errors(&self) {
-        assert!(
-            self.errs.is_empty(),
-            "expected to find no errors, but found: {:?}",
-            self.errs
-        );
-    }
-
-    /// Return all the successfully retrieved directory entries in the order
-    /// in which they were retrieved.
-    pub fn ents(&self) -> &[windows::DirEntry] {
-        &self.ents
-    }
-
-    /// Return all file names from all successfully retrieved directory
-    /// entries.
-    ///
-    /// This does not include file names that correspond to an error.
-    pub fn file_names(&self) -> Vec<OsString> {
-        self.ents.iter().map(|d| d.file_name_os().to_os_string()).collect()
-    }
-
-    /// Return all the successfully retrieved directory entries, sorted
-    /// lexicographically by their file name.
-    pub fn sorted_ents(&self) -> Vec<windows::DirEntry> {
-        let mut ents = self.ents.clone();
-        ents.sort_by(|e1, e2| e1.file_name_u16().cmp(e2.file_name_u16()));
-        ents
-    }
-
-    /// Return all file names from all successfully retrieved directory
-    /// entries, sorted lexicographically.
-    ///
-    /// This does not include file names that correspond to an error.
-    pub fn sorted_file_names(&self) -> Vec<OsString> {
-        self.sorted_ents().into_iter().map(|d| d.into_file_name_os()).collect()
-    }
-}
-
 /// A helper for managing a directory in which to run tests.
 ///
 /// When manipulating paths within this directory, paths are interpreted
@@ -231,56 +115,6 @@ impl Dir {
     {
         let mut results = RecursiveResults { ents: vec![], errs: vec![] };
         for result in it {
-            match result {
-                Ok(ent) => results.ents.push(ent),
-                Err(err) => results.errs.push(err),
-            }
-        }
-        results
-    }
-
-    #[cfg(unix)]
-    pub fn run_unix(&self, udir: &mut unix::Dir) -> UnixResults {
-        let mut results = UnixResults { ents: vec![], errs: vec![] };
-        while let Some(result) = udir.read() {
-            match result {
-                Ok(ent) => results.ents.push(ent),
-                Err(err) => results.errs.push(err),
-            }
-        }
-        results
-    }
-
-    #[cfg(target_os = "linux")]
-    pub fn run_linux(&self, dirfd: &mut unix::DirFd) -> UnixResults {
-        use crate::os::linux::{getdents, DirEntryCursor};
-        use std::os::unix::io::AsRawFd;
-
-        let mut results = UnixResults { ents: vec![], errs: vec![] };
-        let mut cursor = DirEntryCursor::new();
-        loop {
-            match getdents(dirfd.as_raw_fd(), &mut cursor) {
-                Err(err) => {
-                    results.errs.push(err);
-                    break;
-                }
-                Ok(false) => {
-                    break;
-                }
-                Ok(true) => {
-                    while let Some(ent) = cursor.read_unix() {
-                        results.ents.push(ent);
-                    }
-                }
-            }
-        }
-        results
-    }
-
-    #[cfg(windows)]
-    pub fn run_windows(&self, h: &mut windows::FindHandle) -> WindowsResults {
-        let mut results = WindowsResults { ents: vec![], errs: vec![] };
-        while let Some(result) = h.read() {
             match result {
                 Ok(ent) => results.ents.push(ent),
                 Err(err) => results.errs.push(err),
@@ -325,7 +159,29 @@ impl Dir {
         src: P1,
         link_name: P2,
     ) {
-        symlink_file(self.join(src), self.join(link_name)).unwrap()
+        #[cfg(windows)]
+        fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
+            use std::os::windows::fs::symlink_file;
+            symlink_file(src, link_name)
+        }
+
+        #[cfg(unix)]
+        fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
+            use std::os::unix::fs::symlink;
+            symlink(src, link_name)
+        }
+
+        let (src, link_name) = (self.join(src), self.join(link_name));
+        imp(&src, &link_name)
+            .map_err(|e| {
+                err!(
+                    "failed to symlink file {} with target {}: {}",
+                    src.display(),
+                    link_name.display(),
+                    e
+                )
+            })
+            .unwrap()
     }
 
     /// Create a directory symlink to the given src with the given link name.
@@ -334,7 +190,29 @@ impl Dir {
         src: P1,
         link_name: P2,
     ) {
-        symlink_dir(self.join(src), self.join(link_name)).unwrap()
+        #[cfg(windows)]
+        fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
+            use std::os::windows::fs::symlink_dir;
+            symlink_dir(src, link_name)
+        }
+
+        #[cfg(unix)]
+        fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
+            use std::os::unix::fs::symlink;
+            symlink(src, link_name)
+        }
+
+        let (src, link_name) = (self.join(src), self.join(link_name));
+        imp(&src, &link_name)
+            .map_err(|e| {
+                err!(
+                    "failed to symlink directory {} with target {}: {}",
+                    src.display(),
+                    link_name.display(),
+                    e
+                )
+            })
+            .unwrap()
     }
 }
 
@@ -384,35 +262,21 @@ impl TempDir {
     }
 }
 
-/// Test whether file symlinks are believed to work on in this environment.
-///
-/// If they work, then return true, otherwise return false.
+#[cfg(any(unix, windows))]
 pub fn symlink_file_works() -> bool {
     use std::sync::atomic::{AtomicUsize, Ordering};
 
-    // 0 = untried
-    // 1 = works
-    // 2 = does not work
     static WORKS: AtomicUsize = AtomicUsize::new(0);
-
     let status = WORKS.load(Ordering::SeqCst);
     if status != 0 {
         return status == 1;
     }
 
     let tmp = TempDir::new().unwrap();
-    let foo = tmp.path().join("foo");
-    let foolink = tmp.path().join("foo-link");
-    File::create(&foo)
-        .map_err(|e| {
-            err!("error creating file {} for link test: {}", foo.display(), e)
-        })
-        .unwrap();
-    if let Err(_) = symlink_file(&foo, &foolink) {
-        WORKS.store(2, Ordering::SeqCst);
-        return false;
-    }
-    if let Err(_) = fs::read(&foolink) {
+    let file = tmp.path().join("file");
+    let link = tmp.path().join("link");
+    File::create(&file).unwrap();
+    if create_file_symlink(&file, &link).is_err() || fs::read(&link).is_err() {
         WORKS.store(2, Ordering::SeqCst);
         return false;
     }
@@ -420,56 +284,19 @@ pub fn symlink_file_works() -> bool {
     true
 }
 
-/// Create a file symlink to the given src with the given link name.
-fn symlink_file<P1: AsRef<Path>, P2: AsRef<Path>>(
-    src: P1,
-    link_name: P2,
-) -> Result<()> {
-    #[cfg(windows)]
-    fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
-        use std::os::windows::fs::symlink_file;
-        symlink_file(src, link_name)
-    }
-
-    #[cfg(unix)]
-    fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
-        use std::os::unix::fs::symlink;
-        symlink(src, link_name)
-    }
-
-    imp(src.as_ref(), link_name.as_ref()).map_err(|e| {
-        err!(
-            "failed to symlink file {} with target {}: {}",
-            src.as_ref().display(),
-            link_name.as_ref().display(),
-            e
-        )
-    })
+#[cfg(not(any(unix, windows)))]
+pub fn symlink_file_works() -> bool {
+    false
 }
 
-/// Create a directory symlink to the given src with the given link name.
-fn symlink_dir<P1: AsRef<Path>, P2: AsRef<Path>>(
-    src: P1,
-    link_name: P2,
-) -> Result<()> {
-    #[cfg(windows)]
-    fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
-        use std::os::windows::fs::symlink_dir;
-        symlink_dir(src, link_name)
-    }
-
+#[cfg(any(unix, windows))]
+fn create_file_symlink(src: &Path, link: &Path) -> io::Result<()> {
     #[cfg(unix)]
-    fn imp(src: &Path, link_name: &Path) -> io::Result<()> {
-        use std::os::unix::fs::symlink;
-        symlink(src, link_name)
+    {
+        std::os::unix::fs::symlink(src, link)
     }
-
-    imp(src.as_ref(), link_name.as_ref()).map_err(|e| {
-        err!(
-            "failed to symlink directory {} with target {}: {}",
-            src.as_ref().display(),
-            link_name.as_ref().display(),
-            e
-        )
-    })
+    #[cfg(windows)]
+    {
+        std::os::windows::fs::symlink_file(src, link)
+    }
 }
